Move the WoA fallback modules into util/woa
Group woa_imshow and woa_librosa under a util/woa package and update
arg_utils and the whisper / gpt-sovits examples accordingly.
