+++ audio_processing/gpt-sovits/example_ailia_voice.py
@@ -4,7 +4,7 @@ import ailia_voice
 
 import os, sys, platform
 if (platform.system(), platform.machine().lower()) in {('Windows', 'arm64'), ('Windows', 'aarch64')}:
-	sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'util'))
+	sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'util', 'woa'))
 	import woa_librosa
 
 import librosa
