+++ audio_processing/whisper/example_ailia_speech.py
@@ -5,7 +5,7 @@ import ailia_speech
 
 import os, sys, platform
 if (platform.system(), platform.machine().lower()) in {('Windows', 'arm64'), ('Windows', 'aarch64')}:
-	sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'util'))
+	sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'util', 'woa'))
 	import woa_librosa
 
 import librosa
