+++ util/woa/__init__.py
@@ -1,0 +1,2 @@
+# Fallback modules for Windows on ARM (WoA), where some binary packages
+# (OpenCV with GUI, librosa, soundfile) are not available.
